+++ instrumentation/jedis/jedis-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v3_0/JedisDbAttributesGetter.java
@@ -56,7 +56,7 @@ class JedisDbAttributesGetter implements DbClientAttributesGetter<JedisRequest, 
     if (!emitStableDatabaseSemconv()) {
       return request.getConnection().getHost();
     }
-    RedisServerTarget target = JedisSingletons.connectionTarget(request.getConnection());
+    RedisServerTarget target = JedisConfiguredTargets.connectionTarget(request.getConnection());
     return target != null ? target.getAddress() : null;
   }
 
@@ -66,7 +66,7 @@ class JedisDbAttributesGetter implements DbClientAttributesGetter<JedisRequest, 
     if (!emitStableDatabaseSemconv()) {
       return request.getConnection().getPort();
     }
-    RedisServerTarget target = JedisSingletons.connectionTarget(request.getConnection());
+    RedisServerTarget target = JedisConfiguredTargets.connectionTarget(request.getConnection());
     return target != null ? target.getPort() : null;
   }
