+++ instrumentation/jedis/jedis-3.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v3_0/ShardedJedisInstrumentation.java
@@ -37,7 +37,7 @@ class ShardedJedisInstrumentation implements TypeInstrumentation {
 
     @Advice.OnMethodEnter(suppress = Throwable.class, inline = false)
     public static Scope onEnter(@Advice.Argument(0) @Nullable List<JedisShardInfo> shards) {
-      return JedisSingletons.openConfiguredTargetScope(JedisServerTargets.ofShards(shards));
+      return JedisConfiguredTargets.openConfiguredTargetScope(JedisServerTargets.ofShards(shards));
     }
 
     @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class, inline = false)
